refactor: 불필요한 코드 제거 및 주석 추가, 패키지 구조 및 메서드 명 변경
#160

#Resolves: #160
#Ref:
