+++ src/main/java/com/sparta/finalproject/domain/air/controller/AirController.java
@@ -1,8 +1,8 @@
-package com.sparta.finalproject.domain.api.controller;
+package com.sparta.finalproject.domain.air.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.sparta.finalproject.domain.api.entity.RealtimeCityAir;
-import com.sparta.finalproject.domain.api.service.AirService;
+import com.sparta.finalproject.domain.air.entity.RealtimeCityAir;
+import com.sparta.finalproject.domain.air.service.AirService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -13,8 +13,9 @@ public class AirController {
 
     private final AirService airService;
 
+    // 미세 먼지 API
     @GetMapping("/air")
-    public RealtimeCityAir getBadair() throws JsonProcessingException {
-        return airService.testAir();
+    public RealtimeCityAir getAirCondition() throws JsonProcessingException {
+        return airService.airConditionGet();
     }
 }
