+++ src/main/java/com/sparta/finalproject/domain/gallery/controller/ImagePostController.java
@@ -16,6 +16,7 @@ public class ImagePostController {
 
     private final ImagePostService imagePostService;
 
+    // 사진 게시글 생성
     @PostMapping("kindergarten/{kindergartenId}/classroom/{classroomId}/gallery")
     public GlobalResponseDto imagePostAdd(@PathVariable Long kindergartenId, @PathVariable Long classroomId,
                                           @ModelAttribute ImagePostRequestDto imagePostRequestDto,
@@ -23,12 +24,14 @@ public class ImagePostController {
         return imagePostService.addImagePost(kindergartenId, classroomId, imagePostRequestDto, userDetails.getUser());
     }
 
+    // 사진 게시글 조회
     @GetMapping("kindergarten/{kindergartenId}/classroom/{classroomId}/gallery/{imagePostId}")
     public GlobalResponseDto imagePostFind(@PathVariable Long kindergartenId, @PathVariable Long classroomId,
                                            @PathVariable Long imagePostId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
         return imagePostService.findImagePost(kindergartenId, imagePostId, userDetails.getUser());
     }
 
+    // 사진 게시글 검색
     @GetMapping("kindergarten/{kindergartenId}/classroom/{classroomId}/gallery")
     public GlobalResponseDto imagePostPageFind(@PathVariable Long kindergartenId, @PathVariable Long classroomId, @AuthenticationPrincipal UserDetailsImpl userDetails,
                                                @RequestParam(required = false, defaultValue = "2000-01-01", value = "start") String start,
@@ -39,14 +42,16 @@ public class ImagePostController {
         return imagePostService.findImagePostPage(kindergartenId, classroomId,userDetails.getUser(),start, end, keyword, page-1, isAsc);
     }
 
-    @PutMapping("kindergarten/{kindergartenId}/classroom/{classroomId}/gallery/{imagePostId}")
-    public GlobalResponseDto imagePostModify(@PathVariable Long kindergartenId, @PathVariable Long classroomId,
-                                             @PathVariable Long imagePostId,
-                                             @ModelAttribute ImagePostRequestDto imagePostRequestDto,
-                                             @AuthenticationPrincipal UserDetailsImpl userDetails) throws IOException{
-        return imagePostService.modifyImagePost(kindergartenId, imagePostId, imagePostRequestDto, userDetails.getUser());
-    }
+//    // 사진 게시글 수정
+//    @PutMapping("kindergarten/{kindergartenId}/classroom/{classroomId}/gallery/{imagePostId}")
+//    public GlobalResponseDto imagePostModify(@PathVariable Long kindergartenId, @PathVariable Long classroomId,
+//                                             @PathVariable Long imagePostId,
+//                                             @ModelAttribute ImagePostRequestDto imagePostRequestDto,
+//                                             @AuthenticationPrincipal UserDetailsImpl userDetails) throws IOException{
+//        return imagePostService.modifyImagePost(kindergartenId, imagePostId, imagePostRequestDto, userDetails.getUser());
+//    }
 
+    // 사진 게시글 삭제
     @DeleteMapping("kindergarten/{kindergartenId}/classroom/{classroomId}/gallery/{imagePostId}")
     public GlobalResponseDto imagePostDelete(@PathVariable Long kindergartenId, @PathVariable Long classroomId,
                                              @PathVariable Long imagePostId,
