+++ src/main/java/com/sparta/finalproject/global/config/jwt/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.sparta.finalproject.domain.jwt;
+package com.sparta.finalproject.global.config.jwt;
 
 
 import com.sparta.finalproject.domain.security.UserDetailsServiceImpl;
