+++ src/main/java/com/sparta/finalproject/global/config/refreshToken/AccessTokenReissuanceController.java
@@ -1,4 +1,4 @@
-package com.sparta.finalproject.domain.jwt;
+package com.sparta.finalproject.global.config.refreshToken;
 
 import com.sparta.finalproject.global.dto.GlobalResponseDto;
 import lombok.RequiredArgsConstructor;
